chore: consolidate models for ease of use
