--- Common/Data/Models/MobEntry.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace PathOfTerraria.Common.Data.Models;
-
-public class MobEntry
-{
-	public float? Scale { get; set; }
-	public string Prefix { get; set; }
-	public decimal Weight { get; set; }
-	public MobStats Stats { get; set; }
-	public List<MobEntryAffix> Affixes { get; set; }
-	public string Requirements { get; set; }
-}
