--- Common/Data/Models/MobEntryAffix.cs
@@ -1,6 +1,0 @@
-﻿namespace PathOfTerraria.Common.Data.Models;
-
-public class MobEntryAffix
-{
-	public string Name { get; set; }
-}
