--- Common/Data/Models/MobStats.cs
@@ -1,7 +1,0 @@
-﻿namespace PathOfTerraria.Common.Data.Models;
-
-public class MobStats
-{
-	public int Level { get; set; }
-	public int Experience { get; set; }
-}
